chore: move some types to api
(There where a few api types that depended on internal types)
