+++ src/api/types.ts
@@ -5,3 +5,13 @@ export type Duration = number
 
 /** Id of a timeline-object */
 export type ObjectId = string
+
+export type InstanceId = `@${string}`
+
+// References:
+
+export type ObjectReference = `#${string}`
+export type ClassReference = `.${string}`
+export type LayerReference = `$${string}`
+export type InstanceReference = `@${InstanceId}`
+export type Reference = ObjectReference | ClassReference | LayerReference | InstanceReference
